Backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cachetools==6.2.4
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.58.1
numpy==1.26.4
openai-whisper==20250625
packaging==25.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.2
zope.interface==8.1.1
